Updated the app so it displays the start and finish points, moved the map logic to a separate controller, fixed minor issues
