+++ GarbageCollectionApp/GarbageCollectionApp/Views/Admin/Index.cshtml
@@ -6,5 +6,5 @@
     <li><a href="/Admin/Citizens">Manage Citizens</a></li>
     <li><a href="/Admin/DisplayBins">Display Garbage Bins</a></li>
     <li><a href="/Admin/CitizenCollections">View Collections</a></li>
-    <li><a href="/Admin/CollectionMap">View Collection Map</a></li>
+    <li><a href="/Map/CollectionMap">View Collection Map</a></li>
 </ul>
